params coef e variavel convertidos de float para int

+++ include/polinomio.h
@@ -11,11 +11,11 @@ struct Polinomio {
 		this->head = nullptr;
 	}
 
-	void inserir(float coef, int expo);
+	void inserir(int coef, int expo);
 	void Simplificar();
 	int G() const;
 	int T() const;
-	float Avaliar(float variavel) const;
+	float Avaliar(int variavel) const;
 	std::string P() const;
 	friend Polinomio operator+(const Polinomio& poli1, const Polinomio& poli2);
 	friend Polinomio operator-(const Polinomio& poli1, const Polinomio& poli2);
